+++ tests/test_no.py
@@ -3,7 +3,8 @@ from django.test import SimpleTestCase
 from django.utils.translation import ugettext_lazy as _
 from django.utils.translation import override
 
-from localflavor.no.forms import NOMunicipalitySelect, NOPhoneNumberField, NOSocialSecurityNumber, NOZipCodeField
+from localflavor.no.forms import (NOBankAccountNumber, NOMunicipalitySelect, NOPhoneNumberField, NOSocialSecurityNumber,
+                                  NOZipCodeField)
 
 
 class NOLocalFlavorTests(SimpleTestCase):
@@ -37,6 +38,39 @@ class NOLocalFlavorTests(SimpleTestCase):
             '+4512345678': error_format,  # wrong country code
         }
         self.assertFieldOutput(NOPhoneNumberField, valid, invalid)
+
+    def test_NOBankAccountNumber(self):
+        error_format = [_('Enter a valid Norwegian bank account number.')]
+        error_checksum = [_('Invalid control digit. Enter a valid Norwegian bank account number.')]
+        error_length = [_('Invalid length. Norwegian bank account numbers are 11 digits long.')]
+
+        # A good source of loads of highly-likely-to-be-valid examples are available at
+        # http://www.skatteetaten.no/no/Person/Skatteoppgjor/Restskatt/Kontonummer-til-skatteoppkreverkontorene/
+        valid = {
+            '7694 05 12057': '76940512057',
+            '7694.05.12057': '76940512057',
+            '7694.05.12057  ': '76940512057',
+            '1111.00.22222': '11110022222',
+            '5555.88.43216': '55558843216',
+            '63450618537': '63450618537',
+            ' 6345.06.20027 ': '63450620027',
+        }
+        invalid = {
+            '76940512056': error_checksum,  # invalid check digit
+            '1111.00.22228': error_checksum,  # invalid check digit
+            'abcdefgh': error_format,  # illegal characters, though it'll fail to create the checksum
+            '1111a00b22222': error_format,  # illegal characters
+            '769405120569': error_length,  # invalid length (and control number for that matter)
+        }
+        self.assertFieldOutput(NOBankAccountNumber, valid, invalid)
+
+    def test_NOBankAccountNumber_formatting(self):
+        form = NOBankAccountNumber()
+        self.assertEqual(form.prepare_value('76940512057'), '7694.05.12057')
+        # In the event there's already empty/blank/null values present.
+        # Any invalid data should be stopped by form.validate, which the above test should take care of.
+        self.assertEqual(form.prepare_value(None), None)
+        self.assertEqual(form.prepare_value(''), '')
 
     def test_NOSocialSecurityNumber(self):
         error_format = [_('Enter a valid Norwegian social security number.')]